QuiniBet/requirements.txt
Flask==2.3.2
python-dotenv==0.19.0
Flask-RESTful==0.3.10
flask_sqlalchemy==3.0.5
flask-jwt-extended==4.5.2
flask-mail==0.9.1
flask_marshmallow==0.15.0
sqlalchemy_filters==0.13.0
Faker==19.3.0
psycopg2==2.9.7
flask-migrate==4.0.4
#request3==2.19.0
bleach==6.1.0